cv_parser_agent/requirements.txt
llama_index
llama_index-llms-gemini
llama_index-embeddings-huggingface
ipykernel
python-dotenv
pydantic
gradio